LayoutTests/sputnik/Conformance/12_Statement/12.14_The_try_Statement/S12.14_A16_T13-expected.txt

CONSOLE MESSAGE: line 80: SyntaxError: Unexpected string "22"
S12.14_A16_T13

PASS Expected parsing failure

TEST COMPLETE
